+++ massa-ledger-exports/src/ledger_entry.rs
@@ -4,8 +4,8 @@
 
 use crate::ledger_changes::LedgerEntryUpdate;
 use crate::types::{Applicable, SetOrDelete};
-use massa_models::amount::{AmountDeserializer, AmountSerializer};
-use massa_models::{Amount, VecU8Deserializer, VecU8Serializer};
+use massa_models::amount::{Amount, AmountDeserializer, AmountSerializer};
+use massa_models::serialization::{VecU8Deserializer, VecU8Serializer};
 use massa_serialization::{
     Deserializer, SerializeError, Serializer, U64VarIntDeserializer, U64VarIntSerializer,
 };
@@ -96,7 +96,7 @@ impl DatastoreDeserializer {
     /// Creates a new `DatastoreDeserializer`
     pub fn new(
         max_datastore_entry_count: u64,
-        max_datastore_key_length: u64,
+        max_datastore_key_length: u8,
         max_datastore_value_length: u64,
     ) -> Self {
         Self {
@@ -106,7 +106,7 @@ impl DatastoreDeserializer {
             ),
             key_deserializer: VecU8Deserializer::new(
                 Included(u64::MIN),
-                Included(max_datastore_key_length),
+                Included(max_datastore_key_length as u64),
             ),
             value_deserializer: VecU8Deserializer::new(
                 Included(u64::MIN),
@@ -230,7 +230,7 @@ impl LedgerEntryDeserializer {
     /// Creates a new `LedgerEntryDeserializer`
     pub fn new(
         max_datastore_entry_count: u64,
-        max_datastore_key_length: u64,
+        max_datastore_key_length: u8,
         max_datastore_value_length: u64,
     ) -> Self {
         Self {
